Drop JSR-305 dependency in favor of org.jetbrains:annotations
We already introduced the org.jetbrains:annotations dependency in d92775c5f3b272b5d3d29de5687ff7c6a34497e0
and we do not need to have two such libraries as dependency.

Since we use it only for annotations, we do not need to have it as transitive dependency.